requirements.txt
selenium==4.8.0
pytest==7.2.0
fpdf==1.7.2
python-dotenv>=1.0.0